Migrated source models to NRML 0.5
