+++ openquake/commonlib/tests/sourcewriter_test.py
@@ -60,11 +60,11 @@ class SourceWriterTestCase(unittest.TestCase):
 class DeepcopyTestCase(unittest.TestCase):
     def test_is_writeable(self):
         parser = SourceModelParser(SourceConverter(50., 1., 10, 0.1, 10.))
-        [sf, cf] = map(copy.deepcopy, parser.parse_sources(ALT_MFDS))
+        groups = map(copy.deepcopy, parser.parse_groups(ALT_MFDS))
         # there are a SimpleFaultSource and a CharacteristicFaultSource
         fd, fname = tempfile.mkstemp(suffix='.xml')
         with os.fdopen(fd, 'wb'):
-            write_source_model(fname, [sf, cf], 'Test Source Model')
+            write_source_model(fname, groups, 'Test Source Model')
         # NB: without Node.__deepcopy__ the serialization would fail
         # with a RuntimeError: maximum recursion depth exceeded while
         # calling a Python object
